Slight Edits
I don't know if using -numpy.inf's are the best way to go, but the example code runs without error. Maybe there are some other's to test it with, I'm just using -numpy.inf's as really small numbers so that any fitness score that is greater than that will make it through to the next gen's. I will update my example script as well.