Method callbacks for on_function and crossover/mutation set to None
The modifications to pygad are annotated as "#barloff" for easy search. This modifications (1) allow to pass the user's Class instance reference for persistence flow of computations; (2) this Class instance reference is now declared in the on_ GA call implementations, set to None as default at the user side, and (3) specifically for on_crossover and on_mutation references now are always executed, even when crossover_type and mutation_type are None.